+++ python/fastdeploy/vision/faceid/contrib/__init__.py
@@ -13,3 +13,5 @@
 # limitations under the License.
 
 from __future__ import absolute_import
+from .insightface import *
+from .adaface import *
