+++ python/fastdeploy/vision/faceid/contrib/adaface/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 from __future__ import absolute_import
-from .contrib import *
+import adaface
