+++ python/fastdeploy/vision/faceid/contrib/insightface/arcface.py
@@ -13,32 +13,31 @@
 # limitations under the License.
 
 from __future__ import absolute_import
-import logging
-from .... import FastDeployModel, ModelFormat
-from .... import c_lib_wrap as C
+from python.fastdeploy import FastDeployModel, ModelFormat
+from python.fastdeploy import c_lib_wrap as C
 
 
-class VPL(FastDeployModel):
+class ArcFace(FastDeployModel):
     def __init__(self,
                  model_file,
                  params_file="",
                  runtime_option=None,
                  model_format=ModelFormat.ONNX):
-        """Load a VPL model exported by InsigtFace.
+        """Load a ArcFace model exported by InsigtFace.
 
-        :param model_file: (str)Path of model file, e.g ./vpl.onnx
+        :param model_file: (str)Path of model file, e.g ./arcface.onnx
         :param params_file: (str)Path of parameters file, e.g yolox/model.pdiparams, if the model_fomat is ModelFormat.ONNX, this param will be ignored, can be set as empty string
         :param runtime_option: (fastdeploy.RuntimeOption)RuntimeOption for inference this model, if it's None, will use the default backend on CPU
         :param model_format: (fastdeploy.ModelForamt)Model format of the loaded model
         """
         # 调用基函数进行backend_option的初始化
         # 初始化后的option保存在self._runtime_option
-        super(VPL, self).__init__(runtime_option)
+        super(ArcFace, self).__init__(runtime_option)
 
-        self._model = C.vision.faceid.VPL(model_file, params_file,
-                                          self._runtime_option, model_format)
+        self._model = C.vision.faceid.ArcFace(
+            model_file, params_file, self._runtime_option, model_format)
         # 通过self.initialized判断整个模型的初始化是否成功
-        assert self.initialized, "VPL initialize failed."
+        assert self.initialized, "ArcFace initialize failed."
 
     def predict(self, input_image):
         """ Predict the face recognition result for an input image
